Created the basic layouts
- Created the basic layout for the project. Since there had to be
changes from last project.
- In StartActivity class, there are two options since we have two
buttons. Start new game and load a game. If new game button is pressed,
it takes the existing format from our RAW resource. Then it is rendered
and the data is extracted from that file.
- I am guessing that there will be the initial file given to us for
serialization. So, this was the purpose.
- Tried using the animation for the coin toss. Didn’t work out. Too
complex.
- Tried using the dialog for the coin toss. Didn’t work out either since
it is too much complex for a simple task.
- Just did the random and displayed the result.
- TO DO: currently, it does not let the user choose heads or tail.
However, it does the random. So, even if the user chooses, there won’t
be any problem.
- Tried the file system for saving the current games. It is kind of
confusing since I don’t know where the file is saved. I searched the
entire storage but didn’t find the file name. However, it runs perfectly
fine.
- Changed the function readData(rawFile) to readData(rawFile, filename,
indicator) where indicator indicates whether to read from raw data or
the filename the user provided.
- Checked in StartPageActivity if the user entered filename already
exists to open. If not, we tell the user about it and do not go through
the entire trouble of reaching up to fileAccess.
- TO DO: write into the file to save the game.
- SOLUTION: While writing, open the board and go through each lines. For
each column, just add a space and append. For each row, add \n so that
we can distinguish later.